Upgrade Jest and Vue CLI, fix tests.
After struggling for a while on this, I managed to figure out that
waitForExpect was essentially broken by the changes to timers in Jest
v27, and that the only viable solution was to replace it with waitFor
from the Vue TestingLibrary library. Other options, like the ones
recommended by Vue Test Utils (`await nextTick()` and
`await flushPromises()`) didn't work with the way the mws library does
things. So, this alternative approach works fine with Jest 27's timer
system, and this is the solution I've decided to go with in order to be
able to upgrade to a new version of Vue CLI.