Resolve type errors and remove unnecessary jank.
Now that the GraphQL types are updated to know that nodes are
non-nullable, the TypeScript types are much more sensible and don't
require us to write the code so defensively. So this removes _a ton_ of
dumb TypeScript jank I've introduced while resolving type errors over
the past day. Hurrah!